requirements.txt
Flask-SQLAlchemy===2.1
Flask-Migrate==1.8.0
mysql-python
